requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.0
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.5
psutil==6.1.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.1
python-multipart==0.0.20
python-socketio==5.12.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.41.3
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.50
ultralytics-thop==2.0.13
urllib3==2.2.3
uvicorn==0.34.0
watchfiles==1.0.3
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
